requirements.txt
Flask==2.3.3
gunicorn==21.2.0
eventlet==0.33.3
python-telegram-bot==20.6
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1
moviepy==1.0.3
numpy==1.24.3
Flask-SocketIO==5.3.6
python-socketio==5.10.0
opencv-python-headless==4.8.0.76
